\name{exec-package}
\Rdversion{1.1}
\alias{exec-package}
\alias{exec}
\docType{package}
\title{Execute scripts in exec directory of packages}
\description{
Support for executing scripts from the exec directory 
of packages
}
\details{
\tabular{ll}{
Package: \tab exec\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2009-08-07\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois <francoisromain@free.fr>

Maintainer: Romain Francois <francoisromain@free.fr>
}
\references{
Writing R extensions 
}
\keyword{ package }
